ContributionRecur - Fix token test failure
This faithfully recreates the hack in the above buildOptions() function, appending _test to the name of test processors,
which allows `CRM_Utils_TokenConsistencyTest::testContributionRecurTokenConsistency` to pass.
But one has to wonder: if we are doing this, why only do it for ContributionRecur, why not for all
option lists containing payment processors?