Afform - Reinstate fillMode param to fix autofill-current-user
There was a bug in autofill-current-user, where any argument in the url would break the functionality.

So this reinstates a param that I've been waffling on. It was added in 12308f3 then deleted in e2c263c
but it turns out to be useful in this context, and also helps optimize the loading of joins,
so let's keep it.