Symfony update 5.4.43
We just updated to 5.4.42 but 3 days ago symfony put out 5.4.43 - seems silly
to upgrade but not to the one that is the latest when we tag

The key PR is
https://github.com/symfony/symfony/pull/57493

And I note that it seems like a sleeper bug depending on server config- ie

https://github.com/symfony/symfony/pull/57493#issuecomment-2242650724
"Ubuntu will join impacted systems on October 10 when 24.10 is released."
The final summary says there are no deprecations but this
comment
https://github.com/symfony/symfony/pull/57493#issuecomment-2188423930
suggests there might be

I don't really know if we have security bundles but it seems like
this fixes a bug that occurs when libxml libraries are used but
it adds a deprecation. Also our composer-installing instances will be
getting this fix so it will be less confusing to patch the tarball

composer update -W symfony/*
Gathering patches for root package.
Loading composer repositories with package information
Updating dependencies
Lock file operations: 0 installs, 3 updates, 0 removals
  - Upgrading symfony/dependency-injection (v5.4.42 => v5.4.43)
  - Upgrading symfony/finder (v5.4.42 => v5.4.43)
  - Upgrading symfony/var-dumper (v5.4.42 => v5.4.43)