(dev/core#4267) Afform - Move registration of scanner service
Overview
---------

This is a twofer. For one, it simplifies code (removes boilerplate).
Additionally, it fixes a highly reproducible variation of the error `You
have requested a non-existent service "afform_scanner".` It was observed
while adding support for user-sync to the installer.

Use-Case
-------

Update the installer to call `CRM_Utils_System::synchronizeUsers()`.

When it gets to this step, it tries to resolve default values for some
token-based fields -- which eventually means that `civi.afform.tokens` is loaded.

Before
------

(1) The service `afform_scanner` is registered in an older way (`hook_civicrm_container`).

(2) The use-case fails with error:

`You have requested a non-existent service "afform_scanner".`

In exploring the backtrace for this exception, we find an oddity:

* The service `civi.afform.tokens` has been instantiated.
* It's trying to access the sibling service `afform_scanner`.
* But that service is unknown/undeclared.

Both services are defined by `ext/afform/core`. Why is one valid while the
other invalid?

Part of the answer is that they have been declared in slightly different ways.

* `civi.afform.tokens` uses `@service` (-working-)
* `afform_scanner` uses the `hook_civicrm_container` (-not working-)

After
-----

(1) The service `afform_scanner` is registered in a newer way (`@servce`) which matches
the other service (`civi.afform.tokens`).

(2) The use-case works.