CliRunnerTest - Fix mistaken assertion
Generally, this test is calling `Civi::paths()->getUrl()` and then fetching the URL.

This test only makes sense if you request an absolute URL. However, as written, it's
actually requesting a relative URL. In the signature for `getUrl()`, we see
a default value `$preferFormat = 'relative'`.

To the extent that this test was passing, it's likely a reflection that the
CLI environment is not properly representative of HTTP environment. It's
starting to fail after some URL bugs in cv-backdrop were fixed.