Mailing Report: call mailingSize only if the queue is empty
This report uses data from the 'queue' tables, except for some
reason the number of recipients, which uses mailingSize(), which
in turn queries the civicrm_mailing_recipients SQL table.

Presumably this is for when the mailing is still in Draft.
Otherwise the data is already loaded from the queue table, so
there is a small performance gain from not querying again.
And while in theory the queue table and the recipients table
should be the same, it could result in inconsistent stats if
someone uses the resendmailing extension, for example.