(dev/core#4462) For page-level auth, prefer to use CiviCRM-only identity.
When authenticating to Civi, you can set the active principal as "Contact" and/or "User".

If you set the user, then the web UI of the CMS is liable to render lots of widgets and
nav-items that are appropriate to logged-in users (like "My Account" or "Logout"). Fixing
this requires site-builder to take extra steps in their CMS.

OTOH, if you don't set the user, then -- in some
configurations/use-cases/add-ons -- you could find some events/listeners
don't run as expected.

This commi makes the default policy to be to -not- set the CMS user (since
that's more likely to work out-of-the-box and matches the approach of older
Profile/Checksum pattern). But you really need to switch, there is a way - tweak the token.

* If the auth-token has a signed claim `userMode=>ignore', then we'll ignore CMS user.
* If the auth-token has a signed claim `userMode=>optional', then we'll load CMS user (if available).
* If the auth-token doesn't explicitly say, then we use a default (`ignore`).