Define generic permission `*authenticated*`
This is similar to the `*always allow*` and `*always deny*` in that it's a
generic permission based on special logic.

For example, suppose you're define a new form "About Me". It should not
be available to the anonymous public, but it should be available
to any logged-in users and/or anyone with authenticated hyperlink.
Then set the required permission to `*authenticated*`.