(Test-Only) Fix MinkBase::login() on WordPress
The `login()` makes a signed link to setup an authenticated session -- and
then requests the link.

If you send the request to the WP backend, then there is a race between two
parties that both want to redirect: `/wp-admin` wants to redirect to the
login-UI, and `authx` wants to authenticate and redirect to your final page.

Currently, authx loses every time.  It always redirects to
`wp-admin/admin.php`.  On a first read of that page, I couldn't recognize a
spot for authx to inject itself early enough to stop this.

This commit does the easy work-around: send the login request to `index.php`
instead. With no preemptive redirect, it's able to proceed to running authx.