SearchKit - Add search templates feature
Search templates are identical to other saved searches, but they cannot be used
as a smart group or placed in an Afform.

They are intended to be cloned and used as the starting place for building a new search.

This adds in the UI a new tab for templates, and a button allowing any existing search
to be saved as a template.

In the database, the new column is_template is added to the civicrm_saved_search table.