CommaKV - Prefer more whitespace
The amount of whitespace was traditionally at the discretion of the person
who fills-in the TEXTAREA for "Product.Options".  However, now that the
values may be unserialized and reserialized, the original whitespace
is filtered/lost.

The whitespace produced by the encoder will be the canonical whitespace.

Cramming it all into one line is OK for the demo data ("White, Black,
Green").  But if you use the suggested "key=value" notation, or if you have
more words, then it gets crampt.