ContributionView: show the contrib id, remove some buttons, layout changes
Also removes the Transaction ID, which is duplicated in the Payment Info (and is specific to the
payment, not the contribution).

The buttons were also removed by cause there are too many of them and
they are redundant.

And finally, the "Payment Summary" was moved up with the LineItems, so that
it looks more like the PDF invoice. Same change applied to ContributionView.