Add legacydedupefinder
=====
Overview
====

This adds a new extension for existing dedupe finder behaviour, allowing us to support the
hook that is jammed in the middle of the dedupe finder code going forwards without
being blocked by it (I have an existing PR that cuts many hours off the
dedupe query but it is bending over backwards to work alongside the legacy
attempts to allow integration).

The new extension is enabled on install and on upgrade and there is no behaviour change unless
a site admin disables it. If they choose to do so then

1) the legacy hook & legacy reserved query wrangling will no longer be called
2) dedupe queries will speed up once the blocked code is also merged
https://github.com/civicrm/civicrm-core/pull/30591