Do not re-render the html in the error for radio fields with option values
This does the same as https://github.com/civicrm/civicrm-packages/pull/416
but in Civi code rather than in packages.

My take on it is that the original sin is that the quickform class for rendering
radio options does not nest them in a div / allow us to add classes to that div
and hence we are working around it by rendering the radio options
individually in the template layer.

If we fix that maybe we revert this - it is mostly a case of
registering a class with more suitable markup in
HTML_QuickForm::registerElementType()