SearchKit - Improve the UI for sql functions
This started out as a bugfix but turned into a UI rewrite.

Before: difficult (sometimes impossible) to change the input type of a sql function value (e.g. text/number/field),
and removing a param could only be done when the type was set to field.

After: Each value has a dropdown to select the input type, and a remove button if allowed.