SearchKit - Refresh DB entities with configurable mode
Overview
---------

This updates the handling of "DB Entity" (TABLE storage) -- which requires some mechanism
to update the content of the table (e.g. the "refresh" operation). This makes the
mechanism configurable, with two options:

* "Truncate and re-fill"
* "Create in background and swap"

Before
------

There's one refresh mechanism: clear the table (TRUNCATE) and re-fill with
regenerated data (INSERT) .

In the period between the truncation and the last insertion, the content of
the table is ill-defined.

After
-----

Choice of refresh modes:

* `truncate` mode -- as before

* `swap` mode -- Make a new table with a temporary name and fill that.
   The original table remains available during the build.  Once ready, swap the
   tables atomicly.

Technical Details
-----------------

Configuration hierarchy allows the mode to be set in two ways:

* The setting `search_kit_entity_refresh` determines the default.
* When calling the `refresh` action, you can override the default, eg

    cv api4 -U admin SK_GrownUps.refresh -v mode=truncate
    cv api4 -U admin SK_GrownUps.refresh -v mode=swap