SearchKit - Refresh DB entities via table-swap
Overview
---------

This updates the handling of "DB Entity" (table-mode). In this mode, you need some mechanism
to update the content of the table (e.g. the "refresh" operation). This updates the mechanism.

Before
------

TRUNCATE the table and re-INSERTs with regenerated data.

In the period between the truncation and regenerated data, the content of the table
is ill-defined.

After
-----

CREATE a new table with a temporary name and fill that. Once ready,
swap the old table and new table (atomicly).