Api4 - Fix weight calculation when moving items to/from a parent group
Fixes bugs in the drag-n-drop functionality of the SearchKit Tree displays.

Before: Weights not adjusted when moving item into a new group;
update fails when moving item to the top-level of the tree.

After: Weights adjust correctly whether moving within or between branches of the tree.

The weight handling code needed updating to not conflate NULL with 0.