Return tempTable name from fillTable
The thresholdQuery function currently accesses a protected property on the DedupeRuleGroup
class - tempTable - by switching fillTable to return this we can move the queries back into the
calling function. There are 3 callers - 1 of which are is code we are cleaning up (which
is not directly affected by this change) & 2 of which are
in code we are separating out & freezing as a legacy path, with
the goal of that code having it's workings 'in class'.

 Within the fillTable query
there are 2 mutually exclusive paths - so access to this protected varible
is the only thing currently uniting the 3. This fixes it to return the
value