dev/core#5720 - Ensure default site email address is respected
A few forms present a dropdown for "From Email Address",
and most of them were not correctly setting the default.
I investigated and found a couple causes:

1. Some forms just didn't bother to set it
2. Some forms *tried* to set it, but it got mangled by the html purifier

For the not-even-trying forms, a pretty simple catch-all solution is to
put the default at the top of the list, thus the change to CRM_Core_BAO_Email::domainEmails

For the ones that were trying and failing, I added the field to
getFieldsToExcludeFromPurification.