FormBuilder - Fix fatal crash with ContactLayout summary blocks
When placing an Afform as a block on the contact summary screen,
the ContactLayout extension crashes. This is due to refactoring done
in 6b25cbda8a21fd2b6f6087696fa5622be7a6efdc which deleted AfformBlock.tpl
in favor of using the generic InlineAfform.tpl everywhere.
Unfortunately this change was not compatible with ContactLayout because
the old template counts on a tpl variable called `$block` whereas the
new template calls it `$afform`.

In lieu of a fatal crash, `$block` is a perfectly fine name for a variable :)