Support apiv4 in test functions individualCreate, organizationCreate, householdCreate
These functions are handy & ubiquitous but when you want to pass
in apiv4 values it gets confusing real fast. This fixes it to
accept version=4 if passed in & to default to apiv4 is all the values
passed in are easily recognized as apiv4 values.

Ideally we can migrate a bunch of places to use the apiv4
although deprecating the apiv3 might be noisy - that's for
another day, removing friction on apiv4 is first