CustomGroup - When deleting an entity, remove references from customGroupExtends
When deleting something that's the criteria for a custom group
(e.g. an activity type, event type, event, financial type, etc)
this will clean up stale references so the custom group no longer
points to entities that don't exist.
If the deleted thing is the last reference, the custom group will
be disabled, as otherwise the empty set would be interpreted as "all".

Fixes dev/core#5749