MixinTestCase - Run all operations the same way. Fix recent test-failure.
A recent update to `cv` reveals an edge-case where these two operations are not always the same:

civicrm_api3('Extension', 'refresh', ...);
$cv->api3('Extension', 'refresh', ...);

This is probably a latent/long-standing bug in `Extension.refresh` (*where it clears some
caches but not others*). However, fixing that it is a bigger thing and not suited the
current urgency.

This change makes the code tidier -- and also sidesteps the
latent/long-standing bug.