CustomField - Refactor quickform addCustomDataFieldsToForm to fetch all fields consistently
Before: Fetches custom fields in 2 different ways and then has to do extra filtering

After: Fetches custom fields in 1 way, no extra filtering needed

TODO: Prior to this refactor, the function only loaded multivalued groups if the entity was Contact.
Keeping that quirk for now because I'm not sure all the ramifications of removing it,
but non-contact forms should be checked to see if it's safe to just remove that condition.