(dev/core#5700) Extensions - "Add New" is supposed to enable
Overview
---------

When using the "Manage Extensions" UI, it presents the "Add New" tab. It has
a list of extensions with an "Install" button -- and (on drilling down) it
clearly says that it will both "Download and install" the extension.

However, after #32285 was merged for 6.2, it subtly changed the behavior
for "Add New" -- downloading but NOT installing.

Before + After
--------------

* (CiviCRM 4.x-6.0): "Add New" performs both download and install.
* (CiviCRM 6.1.0): "Add New" only downloads -- it skips installation. You see
  the new extension is available (but not yet installed).
* (Patch): "Add New" performs both download and install. You see the new
  extension is live.