Api4 - More efficient ACL clause handling
This was old code ported over from api3 and it wasn't well-tuned to working with Api4 explicit joins.
Deduping is important for efficiency so the same clauses don't get added to the sql multiple times.

Before: Deduping only worked for implicit joins

After: Deduping works for any combination of explicit and implicit joins