CiviImport - Remove always false portion of clause
Because this condition was checking the wrong strings
("new" instead of "NEW", "valid" instead of "VALID")
the condition was always false so can just be removed.

It's best to not check the status anyway because this is
an API, conditions like this belong in the WHERE clause,
not in the run function.