CustomGroup - Fix regression where multi-record option never appears
This is a theme-specific bug: on traditional Greenwich, it works fine,
but on Riverlea themes the fields remain hidden.

This is due to different interpretations of the "hiddenElement" class,
to which Riverlea adds `!important` and Greenwich does not.

Why fight it? This fix will work on any theme and is more consistent
with what the javascript does anyway.