(REF) AfformSaveTrait - Extract the StringScanner class
MOTIVATION: Make it easier to write a unit-test that calls the built-in string-scanner.

BEFORE: The `AfformSaveTrait` has several helper functions to perform scanning. The
results are stored in a buffer ($this->stringTranslations).

AFTER: The `AfformSaveTrait` uses the `StringScanner`. This class has all the same
helper functions and a similar buffer.