Update standalone install to be non-destructive
This changes the pre-install such that it augments rather than drops & re-creates any pre-existing
civicrm_uf_match table.

Note this removes the block on installing it onto a drupal site. It is something we are
doing as part of our migration so being able to install this in that context is good.
I did think about putting some other check in there to enforce 'deliberateness'
but I feel like the fact the extension is hidden covers that - and now
that enabling is is non-destructive the consequences are not the same...