JobManager - Fix reporting of APIv3 errors
Overview
--------

When a Scheduled Job fails, we should report the error as an error (consistently).

Before
------

If `civicrm_api()` encounters an error, it -might- report the result to
the `$logger` as an error. Or it might not.

* It will `Exception`s raised by `civicrm_api()`
* It will not report other `Throwable`s.
* It will not report APIv3-style error-arrays (`$result['error_message']`).

After
-----

If `civicrm_api()` encounters any kind of (`Exception` or `RuntimeException`
or `$result['error_message']`), the resulting log-entry will be properly
categorized as an error.