RiverLea - Improve field width handling
This significantly improves the SearchKit UI by setting sensible input widths,
enabling inline-forms to appear much less cluttered.

In Riverlea, !important was added to the width override classes, because they are overrides.
And a new `crm-auto-width` class was added which allows a Select2 to only use as much
width as it needs.