oauth-client - Fix the "Inspect" URL
I suspect this regressed when 48b3a2bbfbd changed the handling of `#` in `CRM.url()`.
(At least, reverting parts of that would fix the "Inspect" link in oauth-client.)
But that's outside the window for _recent_ regressions.

Maybe it's possible to teach `CRM.url()` to handle fragment-query-params.
But it doesn't feel like a common need right now, and modifying `CRM.url()`
is more risky. This mechanism is fairly simple.