Add apiv4 support to Dedupe.getduplicates
Retrieving exisiting DB duplicates (ie deduping the db
rather than checking for specific parameters) has long accepted 'criteria'
in an apiv3 format - which it then filters a bit.

However, this doesn't work for modified_date => ['BETWEEN' =['then', 'now']]
because deep within apiv3 the Query object cannot handle that.

These days I think anything that doesn't work with apiv3 is a hint to make it
work with apiv4 so this updates the call to accept 'where' in that
criteria array & to understand that it is an apiv4 call
if that is passed in