Payment Processor - Admin screen quietly copies live credentials to testing
Overview
--------

Fix a weird behavior in managing the "Payment Processor" records, where the "Live"
credentials are quietly copied for use as "Test" credentials.

Steps to Reproduce
------------------

* Navigate to "Admin > System > Payment Processors"
* Create a new "Payment Processor" of any kind (or edit an existing one)
* Set the live "User Name" to "Bob".
* Set the test "User Name" to "" (blank).
* Save.
* Edit the again.
* Observe: What is the test "User Name"?

Before
-------

Any blank value in the "Test" section is autofilled with data from the "Live" section.

After
-----

Live is live, and test is test, and ne'er the twain shall meet.

Comments
---------

Why does the form do this?  I traced the code to
https://issues.civicrm.org/jira/browse/CRM-18061 (circa 4.7.3=>4.7.4), which
attempted to fix a different save-bug on these fields.  This does _not_
appear to be intentional.

And it shouldn't be intentional -- that would be kinda crazy.  With this
bug, CiviCRM's test payments would (by default, absent other configuration)
be sent for live execution.

There are mitigating factors -- e.g.  for PayPal, the "Site URL" and "API
URL" are prefilled with distinct values.  So the real effect depends on the
specific payment-processor.