Ensure mailing widget correctly handles html characters
This was a strange one because Api4 is *supposed* to transparently handle
encoding/decoding of html characters. But after some digging I realized that
the alias "label" used by the union query is on the list of fields to *not*
decode, see `CRM_Utils_API_HTMLInputCoder::getSkipFields`.

Changing the alias to a non-conflicting name solved the problem, but this entire
escape-on-input system is of course terrible and needs to be ripped out.

Fixes dev/core#4561