Smarty - Remove "bonkers" way of fetching help titles
Popup help text is loaded by the `CRM.help()` function. It takes a fileName and fieldID, and fetches help
text from the given `.hlp` file via ajax. Note that it does *not* fetch the title. The optional title must
be passed into the function.

The smarty `{help}` function generates markup for a link which calls `CRM.help()` on-click.
In order to pass a "title" to that function, it has 3 ways of figuring it out:

1. Explicit: `$title` can be passed directly in like `{help id="first_name" title="First Name"}`.
   Simple but verbose.
2. DRY: If `fieldID` matches the name of a field on the form, it can infer title from the field's textLabel.
   This is the most elegant but alas was broken. As of 33d199dd it was checking a variable named `$vars`
   which had not been defined yet.
3. Bonkers: The last method was very complex and, according to the docblock, "a bit bonkers." It would fetch,
   load, and parse the `.hlp` file, just to get the title which is explicitly written there using the magic
   string concatenation of `-title` onto the $fieldID.

The last method is IMO not worth supporting. It's even more verbose than method 1, and way more complex than
necessary. This PR removes it, and restores method 2 to working order.

The immediate consequence of this PR is that some help popups will lose their titles. IMO that's acceptable,
and can be fixed as needed. Many titles will continue to work now that method 2 is fixed, and titles using
method 1 are unaffected.

Note: This change affects Smarty pages/forms only. JS/Angular callers of `CRM.help` are unaffected.