Make icon links accessible
Font awesome icons were used to directly embed into <a> tags. As these
icons need to have a role="img" to be accessible to users with
custom stylesheets (with custom font-families) separate icons needed
to be added with role=img and aria-hidden attributes.

For screen readers all links have:
- been labelled with aria-labels
- been given the role button as it more accurately reflects their
function