Fix filter on relationship tab
Fixes dev/core#6128

This regressed with commit 739d2e31; the problem was that the RelationshipCache entity wasn't declared as a DAOEntity and therefore SavedSearchInspectorTrait::getQuery failed to load a query object for it.

I've fixed the metadata for that entity and a couple others, and also added a fallback in SavedSearchInspectorTrait::applyFilter so it's not so heavily dependent on the query.