SettingField - Get help from metadata
Help text was previously double-entered: once on the setting metadata (without it, the form wouldn't know there was any help to show) and then again in a .hlp file for the actual contents, even though they were identical to what was in the metadata (or not, it was a mess).

Now the setting metadata is the only place to store `help_text` and it will be escaped as plain text unless provided as `help_markup`. An array of values for `help_text` is also accepted and will be formatted as paragraphs.

In cases where the text generation needs some logic, it can be implemented in a `metadata_callback`.

Followup commits will cleanup the help in setting metadata files and delete the redundant `.hlp` files.