Settings - Fix generic form to allow dynamic additions
It's possible that a setting field's `metadata_callback` will add or remove it from a form. In order to support that, the form must start with fully-resolved metadata for all settings, before applying filters.

Picks out and cleans up some functionality from `CRM_Admin_Form_SettingTrait` to make it more efficient, also renames the default section to 'default' instead of '', which gives it a more sensible css class.