Fix float precision edge case (server dependent)
In some cases the value php holds when it goes to save the Contribution
is different to the value that is actually saved due to precision rounding
This relates to the serialize_precision ini variable but appears to also
have some other factors impacting it.

We have 3 places where we conditionally try to load a contribution once
saved so it is useful to know that it is always retrieved and seems
worth any trade off. On the other side if we retrieve it right after
save we can unset the values that might cause inaccuracies