Extensions - Ensure that submodules are found
Adding support for submodules means that we have a greater effective depth.
But things will get spastic/breaky if we start distributing extensions
with new depths (given current limit of 3).

The original purpose of maxDepth was to avoid broad scans over large trees.
But for submodules, it's not a broad scan -- we have a precise pattern that
runs faster.