Modal - Better handling of js/css resources added in popups
Before: css/js was renderd in the ajax-snippet region as markup. When the dialogue was closed, they would disappear from the DOM, which sometimes produces unexpected results.

After: css/js/settings are sent to client as structured data during ajax requests so they can be properly added to the document.

Fixes dev/core#6187