(dev/core#6200) OAuth - On return, require valid `state` but no special perms.
We already guard the initiation of the OAuth flows -- if you're not allowed
to use the flow, then you won't have a valid `state`, and the page-controller
will reject you.

Relaxing this makes it easier to do situational OAuth-flows -- e.g.
someone with `administer payment processors` can begin a flow for
a specific payproc, and we won't reject them upon return.