Test Extensions - Switch to mgmt:hidden
Overview
--------

Make it easier to introduce test-only extensions for use in end-to-end tests.

This PR changes the mechanism by which we hide test-oriented extensions. In both cases,
test-extensions are generally hidden. But the mechanism is different

Before
------

The folder `tests/` is scanned for extensions -- but only when phpunit is
running (i.e.  only when constant `CIVICRM_TEST` is set).

* Consequently, the scan is operative during _headless_ testing.
* But for E2E testing, `phpunit` isn't part of every process. (You can have
  HTTP requests and other subprocesses.) Consequently, items here are
  ill-defined.

After
-----

The folder `tests/extensions/` is scanned for extensions, unconditionally. But note that:

* The folder `tests/` doesn't exist on most production systems. (Excluded via `distmaker` and `.gitattributes`.)
* The folder `tests/extensions/` is much smaller than `tests/`. (When it exists, it's less expensive to scan.)
* All items in `tests/` are flagged with `mgmt:hidden`, so they don't appear in web UI.