RouterTest - Fix false-negatives on WordPress
This end-to-end test (new in 6.10) emits false-negatives when run in the nightly jobs on
WordPress. This fixes it by tweaking the sample values.

NOTE: The general idea is this: the test could return common status-code like 200,
but that might pass for the wrong reason. To see if the framework really respects
the status-code, you have to put an unusual value.

Before
------

The unusual values (498, 499) are too unusual for WordPress.

Specifically, WordPress's `status_header()` utility wants to find the code in a registry. If
the status-code is unregistered, then WP coerces it to 200.

After
-----

The unusual values (429, 428) are still unusual... but the appear WP's
registry. So it's not coerced to 200.