PSR-7 - Normalize getPath()
Overview
--------

In the daily E2E tests, we see conflicting results for PageWithPsr7Test, e.g.

* It passes on Standalone and Drupal 10.
* It fails on Drupal 9.

This normalizes the behavior to pass on more UF's -- and it improves compliance with PSR-7.

Before
------

* The test for `$request->getPath()` fails on D9 (but passes on SA/D10/WP/BD).
* The value of `$request->getPath()` deviates from PSR-7 (in how it reports special characters in the path).

After
-----

* The test for `$request->getPath()` passes on D9 (and as far as I've tested -- SA/D10/WP).
* The value of `$request->getPath()` reports special characters in the expected way.

Technical Details
-----------------

Why did D9 and D10 disagree? I suspect the difference is Guzzle. (Guzzle
provides the implementation of ServerRequestInterface, but each UF may have
a different version of Guzzle.)

This change *would* be a small BC-break, but we caught it early. The
functionality is new in 6.10.alpha, and this fix is for 6.10.beta.